비슷한 역할 객체 통합
